Rearranging vector FE PA code in different files.
