+++ fem/CMakeLists.txt
@@ -19,6 +19,7 @@ set(SRCS
   bilininteg_divergence.cpp
   bilininteg_hcurl.cpp
   bilininteg_hdiv.cpp
+  bilininteg_vectorfe.cpp
   bilininteg_gradient.cpp
   bilininteg_mass.cpp
   bilininteg_vecdiffusion.cpp
