+++ fem/bilininteg_hdiv.cpp
@@ -23,6 +23,11 @@ using namespace std;
 namespace mfem
 {
 
+// Local maximum size of dofs and quads in 1D
+constexpr int HDIV_MAX_D1D = 5;
+constexpr int HDIV_MAX_Q1D = 6;
+
+
 // PA H(div) Mass Assemble 2D kernel
 void PAHdivSetup2D(const int Q1D,
                    const int NE,
